Move svc code to internal folder and update paths
