+++ services/web-frontend/internal/public/favicon.ico
