+++ services/web-frontend/internal/public/logo192.png
