+++ services/web-frontend/internal/public/logo512.png
