+++ services/web-frontend/internal/public/manifest.json
