+++ services/web-frontend/internal/src/react-app-env.d.ts
