+++ services/web-frontend/internal/src/redux/actions/silos.ts
