+++ services/web-frontend/internal/src/redux/reducers/silos.ts
