+++ services/web-frontend/internal/src/styles/App.css
