+++ services/web-frontend/internal/src/styles/Sidebar.css
