+++ services/web-frontend/internal/src/styles/Topbar.css
