+++ services/web-frontend/internal/src/ui/components/Workspace.tsx
