+++ services/web-frontend/internal/src/ui/containers/Panel.tsx
